Minor renaming and restructuring
DisplayData & DisplayDataStream were confusing names, renamed to more
appropriate DisplayScreen (which handles stuff on the screen), and
ProcessDataStream (which handles the stream to and from the host).

Keyboard & Buffer both capitalized.

Some methods renamed to better suit their purpose, and Socket connection
removed from Keyboard (because Keyboard should only be concerned with
the Keyboard and handling of it), and placed in ProcessDataStream.