Fix Font change and cnct/disc/cnct bug
If the font was changed, it only affected the currently active screen
(default or alternate); moved method to ProcessDisplay and applied font
change to both.

Connect/Disconnect/Connect resulted in a segfault; this was caused by
deleting the SocketConnection variable which is created once on entry to
the application.

NOTE: This will probably need to change when multiple sessions are
supported.