Implement multiple windows and MRU
Implementation of multiple document windows, and a most-recently-used
list.

Changes here had to include scaling which was also thrown up as in error
when scaling from a very small window to a maximised window. This may
need revisiting in the future if other rounding type errors creep in.

It may be necessary to create an artificial size for QGraphicsScene
before scaling it to fit into the view.

It may be necessary to add a 'geometryChange' method so that the
QGraphicsRectItems and QGraphicsLineItems can be repositioned according
to the current size.