Initial clipboard copy implementation
DisplayScreen - store screen address (3270 x,y) in Text item so that
copy command can determine when a newline is needed. Fix size of graphic
window so that we don't have tiny fractions for sizing.

Keyboard - add TerminalView to constructor so that Keyboard cut/copy can
get text from the screen.

TerminalTab  - to pass TerminalView to Keyboard, and to ignore geometry
of view (as we're forcing an articial size for the scene).

TerminalView - implement clipboard processing with rubber band. Force
ignoreaspect ratio.

Text - store space on screen when null (0x00) is used from datastream to
space fill on screen. getEBCDIC() will return 0x00 instead of space in
this instance. Add getX() and getY() to identify where a Text is
positioned on screen.