Fix segfault with FontSelection
When a font was selected, and the style from the previous selection was
not available in the new font, a segfault occurred when "OK" was
pressed.

Rather than make the user choose a new style, default to the first in the
list of the new font.

Add code to pick the next smallest font size, should the previously
selected size not exist in the new font - this hasn't happened during
testing so far.