Move attributes to text class + bug fixes
The attributes for a given cell are now contained within the text class,
thus ensuring all character-related data is together. In addition, this
removes another 'new' and a C-style array.

Also a bug fix for keyboard entry that took the character attribute of a
cell rather than the field attribute.