Move settings into their own class
The preferences dialog class contained both functions to manage the
dialog and to manage settings, along with signals for changes to those
settings.

In order to keep the functionality of the preferences dialog purely to
those required for the dialog itself, all settings are moved into a
separate class.

Additionally, no local variables are used to correctly show the current
settings when the dialog is displayed; rather they are obtained always
from the current settings. (This means nothing has to be reverted if/
when the user presses cancel)