Fix SEGV, font problems and convert setFont to signal
A segfault frequently happened on start-up, when referencing a recently-
used session, caused by referencing primaryScreen & secondaryScreen
before they were initialized.

Reinstated anti-aliasing, kerning, smoothing and hinting for a better
font feel, and resigned to the fact that it's the font metrics that
prevent some fonts from joining up line-draw characters (exhibited when,
for example, displaying a drop-down menu in ISPF).

setFont was incorrectly being called when the font was changed in
Preferences, and now converted to a signal (ticking off a TO-DO!).