Fix insert/delete character attributes
When inserting or deleting characters using the keyboard, the character
attributes in (say) an ISPF Edit session did not move with the character
being moved, leading to the colouring of the characters on the line
changing when they were moved.

Modified the setAttr routine to be more logical in its definition.