Improve keyboard structure and rename TerminalTab
ProcessDataStream had several routines that were only present because
TerminalTab connected the keyboard to ProcessDataStream in order to
process events such as moving the cursor.

ProcessDataStream is intended, as the name suggests, processing the 3270
data stream, and not for handling keyboard events. All routines to do
with the keyboard now removed, and either added to DisplayScreen or
removed entirely as the call from ProcessDataStream simply passed
control to DisplayScreen anyway.

One downside of this is that now the keyboard events must be processed
by the screen that is currently active (primary or alternate) so when
EW/EWA are received, the keyboard is switched between them, potentially
slowing things down.

TerminalTab renamed to Terminal as the 'Tab' part doesn't make any sense
now.