Performance improvements and minor tweaks
A performance improvement to drastically cut time spent searching for
the start of a field when the incoming data stream has few fields in it
(SDSF output on a large screen size for example). Field Starts are now
stored in the Cell so any Cell knows the field it belongs to.

This may also enable further efficiencies in the future for updating a
Cell's attributes.

Fixed a segfault in deleteChar when deleteing a character from a field
that ended on the first display position.

Removed various debugging output.

Updated the dump of the incoming datastream.

Add Docs to .gitignore for now.
Added Doxyfile for use with Doxygen.