Correction to a Start Field overwrite
When a Start Field attribute is overwritten by incoming data (such as
when a WRITE command is used to redraw the screen, rather than ERASE
WRITE) the overwritten field needs to use the attributes from the
previous field on the screen.

The calculation for the start and end positions was incorrect, resulting
in the wrong attributes being used.