Convert to single QGraphicsItem for display
The display matrix was made up of a QGraphicsItem, containing a Cell for
each character; the cell contained a QGraphicsSimpleTextItem and a
QGraphicsLineItem for the underscore. On a 24x80 screen, that means 1920
x 2 (3840) QGraphicsItems, and on a large display 160x80 that's 13440 x
2 (26880 items).

In an effort to improve performance of this and to simplify the code,
this is reduced to a single QGraphicsItem for the display, which now
performs the paint() function internally.

A new status bar also introduced to replace the LockIndicator which only
managed the XSystem part of the status bar.

All of this affected many initialisation paths and routines, but should
have improved display performance and responsiveness. It also has had
the added benefit that font sizes in the font selection dialog are now
honoured, giving a finer control for the user.