Implement font tweaks for zero
Implement the capability for a zero to be drawn with an overlaid dot or
slash - for those fonts that don't already have that. It can be
difficult to distinguish between zero and 'O' on screen in certain
fonts, so this adds a visual overlay so the characters appear
differently on screen.